Inline recorder formatting and add runtime data fields
